requirements.txt
streamlit>=1.20.0
pandas
numpy
reportlab
Pillow
